+++ src/components/Popover/index.stories.js
@@ -1,35 +1,20 @@
-import React, { useRef } from 'react';
-import State from 'util/storybook-docs/State';
+import React from 'react';
 import Button from 'components/Button';
+import State from 'util/storybook-docs/State';
 import Popover from '.';
 
 const Template = (args) => <Popover {...args} />;
 export const Primary = Template.bind({});
 Primary.args = {
-  transitionName: 'GrowFast',
-  trigger: (
-    <div className="margin--bottom">
-      <Button label="Click Me" />
-    </div>
-  ),
-  children: (
-    <div
-      className="padding--all"
-      style={{ outline: '3px dotted red', background: '#FFFFFF' }}
-    >
-      popover content
-    </div>
-  ),
+  triggerElement: <Button label="Click Me" />,
+  children: <div className="padding--all">popover content</div>,
 };
 
 export const styledPopoverContent = Template.bind({});
 styledPopoverContent.args = {
-  trigger: <Button theme="outlined" label="Open popover" />,
-  position: 'bottom',
-  alignment: 'start',
-  distance: 8,
+  triggerElement: <Button theme="outlined" label="Open popover" />,
   children: (
-    <div className="bgColor--white rounded--all elevation--2 padding--all">
+    <div className="bgColor--white rounded--all padding--all--s">
       <h3 className="type--head3">Look at me</h3>
       <p>
         <em>i am the popover</em>
@@ -44,7 +29,8 @@ export const controlledPopover = (args) => (
     render={(myIsOpen, setMyIsOpen) => (
       <Popover
         {...args}
-        trigger={
+        onClickOutside={() => setMyIsOpen(false)}
+        triggerElement={
           <button
             onClick={() => {
               setMyIsOpen(!myIsOpen);
@@ -53,16 +39,9 @@ export const controlledPopover = (args) => (
             Open popover
           </button>
         }
-        position="bottom"
-        alignment="start"
-        interactionMode="controlled"
         isOpen={myIsOpen}
-        onUserDismiss={() => {
-          setMyIsOpen(false);
-        }}
-        distance={8}
       >
-        <div className="bgColor--white rounded--all elevation--2 padding--all">
+        <div className="bgColor--white rounded--all padding--all--s">
           <h3 className="type--head3">Look at me</h3>
           <p>
             <em>i am the popover</em>
@@ -73,98 +52,7 @@ export const controlledPopover = (args) => (
   />
 );
 
-controlledPopover.parameters = {
-  docs: {
-    description: {
-      story:
-        'The "controlled" `interactionMode` follows the conventions of a [controlled component](https://reactjs.org/docs/forms.html#controlled-components), meaning the open state of the popover is exclusively controlled by the prop `isOpen`. To update the state of `isOpen` based on user interaction, you must use the `onUserDismiss` callback. The `onUserDismiss` function is invoked when 1. The user clicks the trigger element 2. The user presses `Escape` 3. the user clicks away from the popover. If you\'d like to disable scrolling of some element when a popover is open you can pass a ref of the scrolling element to the `disableScrollRef` prop.',
-    },
-  },
-};
-
-export const closeOnScroll = (args) => {
-  const containerEl = useRef(null);
-  return (
-    <div
-      ref={containerEl}
-      style={{ height: 200, border: '1px dashed red', overflow: 'scroll' }}
-    >
-      <div style={{ height: 500, background: 'yellow' }}>
-        <Popover
-          {...args}
-          trigger={<Button theme="outlined" label="Open popover" />}
-          position="bottom"
-          alignment="start"
-          distance={8}
-          closeOnScrollRef={containerEl}
-        >
-          <div className="bgColor--white rounded--all elevation--2 padding--all">
-            <h3 className="type--head3">Look at me</h3>
-            <p>
-              <em>i am the popover</em>
-            </p>
-          </div>
-        </Popover>
-      </div>
-    </div>
-  );
-};
-
-closeOnScroll.parameters = {
-  docs: {
-    description: {
-      story:
-        "If you'd like the ability to close the popover once the user scrolls some element, you can pass a ref of that element to the `closeOnScrollRef` prop. If you're not sure how to handle scrolling behavior with a popover, choose this option.",
-    },
-  },
-};
-
-export const disableScrolling = (args) => {
-  const containerEl = useRef(null);
-  return (
-    <div
-      ref={containerEl}
-      style={{ height: 200, border: '1px dashed red', overflow: 'scroll' }}
-    >
-      <div style={{ height: 500, background: 'yellow' }}>
-        <Popover
-          {...args}
-          trigger={<Button theme="outlined" label="Open popover" />}
-          position="bottom"
-          alignment="start"
-          distance={8}
-          disableScrollRef={containerEl}
-        >
-          <div className="bgColor--white rounded--all elevation--2 padding--all">
-            <h3 className="type--head3">Look at me</h3>
-            <p>
-              <em>i am the popover</em>
-            </p>
-          </div>
-        </Popover>
-      </div>
-    </div>
-  );
-};
-
-disableScrolling.parameters = {
-  docs: {
-    description: {
-      story:
-        "If you'd like to disable scrolling of some element when a popover is open you can pass a ref of the scrolling element to the `disableScrollRef` prop.",
-    },
-  },
-};
-
 export default {
   component: Popover,
   title: 'components/Popover',
-  parameters: {
-    docs: {
-      description: {
-        component:
-          'Base popover component. Bring your own trigger and popover content. What this component does: Handle events according to `interactionMode` passed in & handle positioning of popover content What this component does **not** do: Content styling (the popover container is just a transparent div) & Arrows. They arent supported, but we can add them in the future. When setting `position` and `alignment` props, the positioning engine will prefer the props you provide, but may move the popover content to avoid edges of the viewport. This behavior is intentional.',
-      },
-    },
-  },
 };
